When using the gem externally, outside the specs, an IntuitRequestException will error with a:
`NoMethodError: undefined method `detail' for #<IntuitRequestException:0x007f96164af8b8>`

To reproduce this the Quickeebooks gem also included,
which is going to be a common scenario.

In short what is happening is that the IntuitRequestException from the
Quickeebooks class is taking over since it is on the global namespace.
That is why it doesn't recognize the `detail` method.

The fix is to put the exception handlers under the library's namespace.

Note: ClassUtili, only being used for conditionally checking of old
exception code has been now been removed.