Dont set Item#rate_percent under the hood
This can cause unexpected results. For example: the Item#unit_price
would not persist in quickbooks as expected because the gem would send a
request with RatePercent: 0.0. That would affect the unit_price setting
it to nothing (0)

Not 100% sure on this but I believe all BigDecimal attributes could take
advantage of this to_xml_big_decimal method introduced here