The ROXML library handles boolean XML true false fields http://roxml.rubyforge.org/rdoc/classes/ROXML/ClassMethods/Declarations.html However, in the code base there is a lot of duplicate predicate methods that could be eliminated by simply adding a '?' to the end of xml_accessor e.g. by setting: xml_accessor :active?, :from => 'Active' this would eliminate: def active?   active.to_s == 'true' end
This PR would possibly break code that relies on an unset attribute to
be false e.g.
Currently, if active is not set it will return false if the predicate
method from above has been created.
After the PR it will return nil.
Moreover, after this PR there is no longer a getter method e.g. 'model.active'.
Currently, a call to 'model.active' would return the string "false" or "true". After the PR
it will be an undefined method.