Relax oauth gem dependency (#449)
The current version constraint requires every app using this gem to also use `oauth` version 0.4.7. For apps that already use `oauth`, as is the case with Clio, bundler fails to resolve a version of `oauth` that works satisfies all conditions.

In this commit, I’ve relaxed the version constraint to allow anything >= 0.4.5 but < 0.5. To get around this constraint, we’ve been using a [fork](https://github.com/nathan-mots/quickbooks-ruby) but I’d love to get back to using the proper gem.