Loosen roxml version constraint
Previously the quickbooks-ruby gem declared a version constraint for
the roxml gem of exactly '4.0.0'. This means that applications using
quickbooks-ruby are unable to use a later version of roxml (such as the
recently released roxml 4.1.0 [1]). To fix this problem, I've loosened
the roxml version constraint to '~> 4.0', which makes it compatible with
any minor versions of 4.x. I didn't choose '>= 4.0' because it's
possible that roxml would release a 5.x version with a breaking change
that would not be compatible with quickbooks-ruby. But as long as roxml
follows semver, it should be safe to change this to '~> 4.0'

[1]: https://rubygems.org/gems/roxml/versions/4.1.0