Remove note about roxml dependency issue
This note was added in 1246b93b248f8691ed1dd06ab1af5076d92051ec because
at the time, roxml hadn't released a new version that removed its use of
the `BigDecimal.new` method (which was deprecated in ruby 2.6 and
removed in ruby 2.7). However, a new version has been released:
https://rubygems.org/gems/roxml/versions/4.1.0
so we can remove this note.