Loosen roxml version constraint (#524)
* Loosen roxml version constraint

Previously the quickbooks-ruby gem declared a version constraint for
the roxml gem of exactly '4.0.0'. This means that applications using
quickbooks-ruby are unable to use a later version of roxml (such as the
recently released roxml 4.1.0 [1]). To fix this problem, I've loosened
the roxml version constraint to '~> 4.0', which makes it compatible with
any minor versions of 4.x. I didn't choose '>= 4.0' because it's
possible that roxml would release a 5.x version with a breaking change
that would not be compatible with quickbooks-ruby. But as long as roxml
follows semver, it should be safe to change this to '~> 4.0'

[1]: https://rubygems.org/gems/roxml/versions/4.1.0

* Remove note about roxml dependency issue

This note was added in 1246b93b248f8691ed1dd06ab1af5076d92051ec because
at the time, roxml hadn't released a new version that removed its use of
the `BigDecimal.new` method (which was deprecated in ruby 2.6 and
removed in ruby 2.7). However, a new version has been released:
https://rubygems.org/gems/roxml/versions/4.1.0
so we can remove this note.

* Configure TravisCI to test against ruby 2.6 and ruby 2.7